feat: update ingestion metrics API return type and move bulk processing of metrics to the backend
- API will now return {serviceName: []IngestionMetrics} where timestamps are already sorted in ascending order by clickhouse
- ErrorsCount will not be returned in ingestion metrics since ingestion metrics only cares about number of logs per second, not errors per second
- update API route from /api/metrics to /api/ingestion-metrics
- rename variables and functions from LogMetrics to IngestionMetrics
- update LogsCount datatype from uint32 to uint64 to store larger nuimber of logs per second

+++ backend/tests/integration/testdata/init-db.sql
@@ -34,8 +34,8 @@ ORDER BY (ServiceName, Timestamp, SeverityNumber);
 CREATE TABLE IF NOT EXISTS logarithm.metrics (
     Timestamp DateTime('UTC'),
     ServiceName LowCardinality(String),
-    LogsCount UInt32, -- stores total number of logs received in 1 second
-    ErrorsCount UInt32 -- stores total number of errors received in 1 second
+    LogsCount UInt64, -- stores total number of logs received in 1 second
+    ErrorsCount UInt64 -- stores total number of errors received in 1 second
 )
 ENGINE = SummingMergeTree()
 PARTITION BY toStartOfHour(Timestamp)
